refactor: move source span exception logging out from logging interface.
