+++ tools/serverpod_cli/lib/src/logger/logger.dart
@@ -1,5 +1,3 @@
-import 'package:source_span/source_span.dart';
-
 /// Serverpods internal logger interface.
 /// All logging output should go through this interface.
 /// The purpose is to simplify implementing and switching out concrete logger
@@ -47,14 +45,6 @@ abstract class Logger {
     bool newParagraph,
     StackTrace? stackTrace,
     LogType type,
-  });
-
-  /// Display a [SourceSpanException] to the user.
-  /// Commands should use this to log [SourceSpanException] with
-  /// enhanced highlighting if possible.
-  void sourceSpanException(
-    SourceSpanException sourceSpan, {
-    bool newParagraph,
   });
 
   /// Display a progress message on [LogLevel.info] while running [runner]
