+++ tools/serverpod_cli/lib/src/util/serverpod_cli_logger.dart
@@ -1,7 +1,10 @@
 import 'dart:io';
 
+import 'package:serverpod_cli/analyzer.dart';
+import 'package:serverpod_cli/src/analyzer/code_analysis_collector.dart';
 import 'package:serverpod_cli/src/logger/logger.dart';
 import 'package:serverpod_cli/src/logger/loggers/std_out_logger.dart';
+import 'package:serverpod_cli/src/util/ansi_style.dart';
 
 /// Singleton instance of logger.
 Logger? _logger;
@@ -41,4 +44,66 @@ Logger get log {
   }
 
   return _logger!;
+}
+
+extension SourceSpanExceptionLogger on Logger {
+  /// Display a [SourceSpanException] to the user.
+  /// Commands should use this to log [SourceSpanException] with
+  /// enhanced highlighting if possible.
+  void sourceSpanException(
+    SourceSpanException sourceSpan, {
+    bool newParagraph = false,
+  }) {
+    var logLevel = LogLevel.error;
+    bool isHint = false;
+
+    if (sourceSpan is SourceSpanSeverityException) {
+      var severity = sourceSpan.severity;
+      isHint = severity == SourceSpanSeverity.hint;
+      logLevel = _SeveritySpanHelpers.severityToLogLevel(severity);
+    }
+
+    if (!(logLevel.index >= log.logLevel.index)) return;
+
+    var highlightAnsiCode =
+        _SeveritySpanHelpers.highlightAnsiCode(logLevel, isHint);
+    var message = sourceSpan.toString(color: highlightAnsiCode);
+
+    write(message, logLevel, newParagraph: newParagraph);
+  }
+}
+
+abstract class _SeveritySpanHelpers {
+  static LogLevel severityToLogLevel(SourceSpanSeverity severity) {
+    switch (severity) {
+      case SourceSpanSeverity.error:
+        return LogLevel.error;
+      case SourceSpanSeverity.warning:
+        return LogLevel.warning;
+      case SourceSpanSeverity.info:
+      case SourceSpanSeverity.hint:
+        return LogLevel.info;
+    }
+  }
+
+  static String highlightAnsiCode(LogLevel severity, bool isHint) {
+    if (severity == LogLevel.info && isHint) {
+      return AnsiStyle.cyan.ansiCode;
+    }
+
+    switch (severity) {
+      case LogLevel.nothing:
+        assert(severity != LogLevel.nothing,
+            'Log level nothing should never be used for a log message');
+        return AnsiStyle.terminalDefault.ansiCode;
+      case LogLevel.error:
+        return AnsiStyle.red.ansiCode;
+      case LogLevel.warning:
+        return AnsiStyle.yellow.ansiCode;
+      case LogLevel.info:
+        return AnsiStyle.blue.ansiCode;
+      case LogLevel.debug:
+        return AnsiStyle.cyan.ansiCode;
+    }
+  }
 }
